Confusing error message for un-imported constant
The confusing part is that the "Did you mean" literally suggests a no-op
edit for the did you mean (why? because internally those two constants
are actually different to Sorbet, but there is a prefix of both that is
being hidden from printing in error messages).

In this case, the outer package imports the nested `::Data` package. The
`::Data` package has no imports. It then references a constant that
comes from the outer package. The confusing error message **only**
happens in this case, not the other way around (if you try flipping
things around, we get a nice "maybe you should import this thing"
error).

You might be tempted to think that this is contrived, but it's the sort
of error that comes up quite easily when either or both of the following
are true:

- You're new to Ruby and/or Stripe's packaging system
- You're working in a part of the codebase where you're not yet familiar
  with what imports (and/or is imported by) what.

We should take this opportunity to show a nice error message to the
user, including any of the following:

- Explicitly say that `Opus::ConsumerAuth` is not imported by
  `Opus::ConsumerAuth::Data`, which is where the similarly named
  constant appears.

- Suggest importing `Opus::ConsumerAuth` into `::Data`.

- Don't suggest importing `Opus::ConsumerAuth` into `::Data`, but offer
  an explanation about how this would introduce a cycle or something.

- Suggest moving the definition of `IdentifierType` into
  `Opus::ConsumerAuth::Data`, so that it can be depended on by both
  packages without introducing cycles.