feat: Add override_documentation for type changes.
It is not fully correct yet, because in some cases we emit
the SymbolInfo.documentation with wrong type information, and
subsequently add override_documentation with the right type
information.

For such situations, we should overwrite the original docs
instead or not emit it in the first place.

I will fix that in a subsequent PR.