fix: Propagate Locs instead of LocOffsets.
Fixes a crash when indexing Homebrew, where we were running into
situations where we'd end up having a LocOffset for a nameLoc,
and it'd be unclear which file it corresponded to. Trying to
recover that information after the fact, such as from the method's
Loc field, seemed unreliable, so I've chosen to carry a Loc instead.