fix: Remove incorrect diagnostic for unknown field (#186)
I'm not exactly sure why, but the result in unresolvedFields (after name
resolution is complete) is non-deterministic. This means that a field may
show up as unresolved in the class in one run, and resolved in another run.

Thus, if the class K is present as a key in unresolvedFields but doesn't
have a corresponding value for a field v, it may just be an artifact of the run.
As such, we can continue with the normal path of name lookups, rather
than having a fast path for this edge case, with a potentially misleading
diagnostic.