fix: Emit reference relationships for constant aliases (#225)
When you have "constant aliases" of the form:

module A
   X = 10
end

module B
  Y = X
end

It would be nice to merge code navigation results for X and Y so
that when you do Find references across either of them, you see
all references for either X or Y.

Part of the logic for this lives in the Sourcegraph backend,
and that needs some fixes in terms of building an extra
inverted index etc. So with this patch, current versions of
Sourcegraph (as of Oct 28 2024) will show merged
references when doing Find references on Y (but not X).

However, this fixes the indexer side of the logic by emitting
Relationships with `is_reference` set, so that the backend
can utilize it.