fix: Correctly emit references for Command#call (#232)
Fixes GRAPH-895.

There are two ways to fix this issue, by emitting a definition
for the synthetic class method generated by Sorbet in rewriter/Command.cc,
as well as emitting a reference for that; OR merely changing
the reference emitting logic to also have a special case
(similar to rewriter/Command.cc). I've picked the second option.

As of now, I don't understand which part of the Sorbet logic
is responsible for making sure that the synthetic method does
not show up when using findMethodTransitive which we're
using for name lookup. So it makes sense to make the simpler
change here.